pyproject.toml
[tool.poetry]
name = "pnwapi"
version = "0.0.0" # Placeholder. Version is set by GitHub releases, and automatically deployed to PyPI.
description = ""
authors = ["Christian <59421913+Cikmo@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/Cikmo/pnwapi"
repository = "https://github.com/Cikmo/pnwapi"
documentation = "https://pnwapi.readthedocs.io/en/latest/"
keywords = ["api", "pnw", "politics and war", "pnwkit", "wrapper"]

[tool.poetry.dependencies]
python = "^3.10"
pnwkit-py = "^2.5.9"
asyncpg = "^0.26.0"
tortoise-orm = "^0.19.2"
numba = "^0.56.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
python-dotenv = "^0.21.0"
pytest-asyncio = "^0.19.0"
pytest-dotenv = "^0.5.2"
pympler = "^1.0.1"
ipykernel = "^6.16.1"
black = {version = "^22.10.0", allow-prereleases = true}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.3.0"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.0.0"

[tool.pyright]
reportImportCycles=false

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
filterwarnings = "ignore:There is no current event loop:DeprecationWarning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
